+++ lib/bottom_sheet/multi_select_bottom_sheet.dart
@@ -52,6 +52,27 @@ class MultiSelectBottomSheet<V> extends StatefulWidget
   /// It will either set the chip color, or the checkbox color depending on the list type.
   final Color Function(V) colorator;
 
+  /// The color of the chip body if the `listType` is set to MultiSelectListType.CHIP
+  final Color chipColor;
+
+  /// Icon button that shows the search field.
+  final Icon searchIcon;
+
+  /// Icon button that hides the search field
+  final Icon closeSearchIcon;
+
+  /// Style the text on the chips or list tiles.
+  final TextStyle itemsTextStyle;
+
+  /// Set the opacity of the chip body. Default is 0.35.
+  final double chipOpacity;
+
+  /// Style the search text.
+  final TextStyle searchTextStyle;
+
+  /// Style the search hint.
+  final TextStyle searchHintStyle;
+
   MultiSelectBottomSheet({
     @required this.items,
     @required this.initialValue,
@@ -68,6 +89,13 @@ class MultiSelectBottomSheet<V> extends StatefulWidget
     this.maxChildSize,
     this.searchPlaceholder,
     this.colorator,
+    this.chipColor,
+    this.searchIcon,
+    this.closeSearchIcon,
+    this.itemsTextStyle,
+    this.chipOpacity,
+    this.searchTextStyle,
+    this.searchHintStyle,
   });
 
   @override
@@ -96,7 +124,10 @@ class _MultiSelectBottomSheetState<V> extends State<MultiSelectBottomSheet<V>> {
       activeColor: widget.colorator != null
           ? widget.colorator(item.value) ?? widget.selectedColor
           : widget.selectedColor,
-      title: Text(item.label),
+      title: Text(
+        item.label,
+        style: widget.itemsTextStyle,
+      ),
       controlAffinity: ListTileControlAffinity.leading,
       onChanged: (checked) {
         setState(() {
@@ -115,22 +146,36 @@ class _MultiSelectBottomSheetState<V> extends State<MultiSelectBottomSheet<V>> {
     return Container(
       padding: const EdgeInsets.all(2.0),
       child: ChoiceChip(
-        selectedColor:
-            widget.colorator != null && widget.colorator(item.value) != null
-                ? widget.colorator(item.value).withOpacity(0.3)
-                : widget.selectedColor != null
-                    ? widget.selectedColor.withOpacity(0.3)
-                    : Theme.of(context).primaryColor.withOpacity(0.3),
+        backgroundColor: widget.chipColor,
+        selectedColor: widget.colorator != null &&
+                widget.colorator(item.value) != null
+            ? widget
+                .colorator(item.value)
+                .withOpacity(widget.chipOpacity ?? 0.35)
+            : widget.selectedColor != null
+                ? widget.selectedColor.withOpacity(widget.chipOpacity ?? 0.35)
+                : Theme.of(context)
+                    .primaryColor
+                    .withOpacity(widget.chipOpacity ?? 0.35),
         label: Text(
           item.label,
           style: _selectedValues.contains(item.value)
               ? TextStyle(
                   color: widget.colorator != null &&
                           widget.colorator(item.value) != null
-                      ? widget.colorator(item.value)
-                      : widget.selectedColor,
+                      ? widget.itemsTextStyle != null
+                          ? widget.itemsTextStyle.color ??
+                              widget.colorator(item.value)
+                          : widget.colorator(item.value)
+                      : widget.itemsTextStyle != null
+                          ? widget.itemsTextStyle.color ?? widget.selectedColor
+                          : widget.selectedColor ??
+                              Theme.of(context).primaryColor,
+                  fontSize: widget.itemsTextStyle != null
+                      ? widget.itemsTextStyle.fontSize
+                      : null,
                 )
-              : null,
+              : widget.itemsTextStyle,
         ),
         selected: _selectedValues.contains(item.value),
         onSelected: (checked) {
@@ -169,7 +214,9 @@ class _MultiSelectBottomSheetState<V> extends State<MultiSelectBottomSheet<V>> {
                             child: Container(
                               padding: EdgeInsets.only(left: 10),
                               child: TextField(
+                                style: widget.searchTextStyle,
                                 decoration: InputDecoration(
+                                  hintStyle: widget.searchHintStyle,
                                   hintText:
                                       widget.searchPlaceholder ?? "Search",
                                   focusedBorder: UnderlineInputBorder(
@@ -192,8 +239,13 @@ class _MultiSelectBottomSheetState<V> extends State<MultiSelectBottomSheet<V>> {
                             child: widget.title != null
                                 ? Text(
                                     widget.title.data,
-                                    style: widget.title.style ??
-                                        TextStyle(fontSize: 18),
+                                    style: TextStyle(
+                                        color: widget.title.style != null
+                                            ? widget.title.style.color
+                                            : null,
+                                        fontSize: widget.title.style != null
+                                            ? widget.title.style.fontSize ?? 18
+                                            : 18),
                                   )
                                 : Text(
                                     "Select",
@@ -203,8 +255,8 @@ class _MultiSelectBottomSheetState<V> extends State<MultiSelectBottomSheet<V>> {
                     widget.searchable != null && widget.searchable
                         ? IconButton(
                             icon: _showSearch
-                                ? Icon(Icons.close)
-                                : Icon(Icons.search),
+                                ? widget.closeSearchIcon ?? Icon(Icons.close)
+                                : widget.searchIcon ?? Icon(Icons.search),
                             onPressed: () {
                               setState(() {
                                 _showSearch = widget.onSearchTap(_showSearch);